feat: fire mapRequestToAsset for all requests if explicitly defined
This restores the ability to run `mapRequestToAsset` on any requests, even if that specific asset already exists in the `ASSET_MANIFEST`.

So for example if you had an `image.png`, `image.webp` and `image.avif` in your manifest, and wanted to dynamically serve the `webp` or `avif` to supported clients when they requested the `image.png`, there's was no previously no simple way to do this. This commit restores this functionality and will always call a user defined `mapRequestToAsset` if set.

See https://github.com/cloudflare/kv-asset-handler/issues/158 for more info