Add support for isolated storage
This change adds an `isolatedStorage` pool option. If enabled, each
test gets its own isolated storage environment. Storage writes in a
test are automatically undone at the end of the test. The storage
environment is copied from the parent suite, ensuring writes from
`beforeAll()`s show up in tests.

This is implemented with an on-disk stack of `.sqlite` files. Before
a suite/test attempt begins, the current storage state is "pushed".
When the suite/test attempt completes/fails, the state is "popped",
undoing all changes. Sticky blobs are used to ensure deleting records
in a test keeps the corresponding blobs around, so they can be used
when the state is "popped".

Closes DEVX-1055